+++ app/services/forest_liana/query_stat_getter.rb
@@ -1,4 +1,17 @@
 module ForestLiana
+  class DatabaseResultAdapter
+    def self.perform(data)
+      case ActiveRecord::Base.connection.adapter_name
+        # NOTICE: MySQL returns a different type for the result when using raw
+        # database request.
+      when 'Mysql', 'Mysql2'
+        { 'value' => data.first.first }
+      else
+        data.first
+      end
+    end
+  end
+
   class QueryStatGetter
     QUERY_SELECT = /\ASELECT\s.*FROM\s.*\z/im
 
@@ -22,7 +35,7 @@ module ForestLiana
       case @params['type']
       when 'Value'
         if result.count
-          result_line = result.first
+          result_line = DatabaseResultAdapter.perform(result)
           if result_line
             if !result_line.key?('value')
               raise error_message(result_line, "'value'")