+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Fixed
+- Live Query - Fix the Live query mode when using MySQL.
 
 ## RELEASE 2.5.4 - 2018-05-14
 ### Fixed